move resize function to file, call from main
