+++ src/resize.js
@@ -1,0 +1,16 @@
+import gameConfig from './gameConfig';
+
+export default function() {
+  var canvas = document.querySelector('canvas');
+  var windowWidth = window.innerWidth;
+  var windowHeight = window.innerHeight;
+  var windowRatio = windowWidth / windowHeight;
+  var gameRatio = gameConfig.width / gameConfig.height;
+  if (windowRatio < gameRatio) {
+    canvas.style.width = windowWidth + 'px';
+    canvas.style.height = windowWidth / gameRatio + 'px';
+  } else {
+    canvas.style.width = windowHeight * gameRatio + 'px';
+    canvas.style.height = windowHeight + 'px';
+  }
+}
